requirements.txt
starkbank
pydantic
python-dateutil
Flask==2.3.3
requests==2.31.0
Flask-Cors==4.0.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
